# openapi.yaml
...
openapi: 3.1.0
info:
  title: Nexus API
  version: 0.1.0
  description: "Generated from the live Fastify route table. Every path carries a
    maturity tag: `durable` — Dedicated handler over durable storage. Safe to
    build on. `dedicated-volatile` — Dedicated handler, but its state lives in
    process memory and is lost on restart. `bridge` — Served by the legacy
    in-memory bridge (routes/api-bridge.ts). Treat its data as synthetic."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/ingest/events/{eventId}:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: eventId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/ingest/signals/{signalId}:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: signalId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/secrets/{name}:
    put:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
      responses:
        "200":
          description: Default Response
...
  /api/v1/secrets/requests/{id}:
    delete:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
  /api/v1/sft/conversations/{id}:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/sft/stats:
    get:
      tags:
        - durable
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
security:
  - bearerAuth: []
...